44. Modulation Strategies Motor Drive/requirements.txt
aesim.simba
numpy
matplotlib
os